Use retry with delay
